refactor: move PrepareCommand hardcoded strings to templates
