+++ lib/src/utils/template_helper.dart
@@ -1,0 +1,9 @@
+class TemplateHelper {
+  static String generate(String template, Map<String, String> variables) {
+    String result = template;
+    variables.forEach((key, value) {
+      result = result.replaceAll('{{$key}}', value);
+    });
+    return result;
+  }
+}
